% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iBAQ}
\alias{iBAQ}
\title{data frame containing iBAQ values}
\format{A data frame with 18889 rows and 33 variables:
\describe{
  \item{MCCTT1}{iBAQ values}
  \item{MCCTT2}{iBAQ values}
  \item{MCCTT3}{iBAQ values}
  \item{MCCT1}{iBAQ values}
  \item{MCCT2}{iBAQ values}
  \item{MCCT3}{iBAQ values}
  \item{MC1}{iBAQ values}
  \item{MC2}{iBAQ values}
  \item{MC3}{iBAQ values}
  \item{MCT1}{iBAQ values}
  \item{MCT2}{iBAQ values}
  \item{MCT3}{iBAQ values}
  \item{MTT1}{iBAQ values}
  \item{MTT2}{iBAQ values}
  \item{MTT3}{iBAQ values}
  \item{MT1}{iBAQ values}
  \item{MT2}{iBAQ values}
  \item{MT3}{iBAQ values}
  \item{sN1}{iBAQ values}
  \item{sN2}{iBAQ values}
  \item{sN3}{iBAQ values}
  \item{sNT1}{iBAQ values}
  \item{sNT2}{iBAQ values}
  \item{sNT3}{iBAQ values}
  \item{iN1}{iBAQ values}
  \item{iN2}{iBAQ values}
  \item{iN3}{iBAQ values}
  \item{iNT1}{iBAQ values}
  \item{iNT2}{iBAQ values}
  \item{iNT3}{iBAQ values}
  \item{pro}{FASTA header}
  \item{E}{Evidence groups}
  \item{FDR}{False Discovery Rate}
  ...
}}
\usage{
iBAQ
}
\description{
A dataset containing iBAQ values for identified proteins
}
\keyword{datasets}

