% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoFRA.R
\name{plot.CompleteEnrichment}
\alias{plot.CompleteEnrichment}
\title{This function plot a heatmap to summarize the results from complete functional enrichment analysis}
\usage{
\method{plot}{CompleteEnrichment}(x, ...)
}
\arguments{
\item{x}{object from complete functional enrichment analysis}

\item{...}{list of additional arguments}
}
\description{
This function plot a heatmap to summarize the results from complete functional enrichment analysis
}
\keyword{"MCT","iNT","sNT","AllT,MCCTT,MTT,MCT,iNT,sNT"))}
\keyword{"MTT","MT","MT","MT","sN","sN","sN","sNT","sNT","sNT","iN","iN","iN","iNT","iNT","iNT"))}
\keyword{#}
\keyword{CC1=CoFRA::completeFunctionalRegulationAnalysis(iBAQ,Func[100:200],Fac,dfComp,NbackGround=142140)}
\keyword{CoFRA::HeatMapEnrichment(CC1,"CC")}
\keyword{Fac=factor(c("MCCTT","MCCTT","MCCTT","MCCT","MCCT","MCCT","MC","MC","MC","MCT","MCT","MCT","MTT","MTT",}
\keyword{Func=CoFRA::getFunctionalCategories("CC")}
\keyword{R}
\keyword{Rstudio}
\keyword{any}
\keyword{because}
\keyword{below}
\keyword{check}
\keyword{commands}
\keyword{data(iBAQ)}
\keyword{dev.off()}
\keyword{device}
\keyword{dfComp=data.frame(Con=c("MCCT","MT","MC","iN","sN","AllC,MCCT,MT,MC,iN,sN"),Tre=c("MCCTT","MTT",}
\keyword{don't}
\keyword{fails}
\keyword{file}
\keyword{files}
\keyword{following}
\keyword{getwd()}
\keyword{graphic}
\keyword{head(str(Func))}
\keyword{heatmap}
\keyword{if}
\keyword{inside}
\keyword{is}
\keyword{library(CoFRA)}
\keyword{note}
\keyword{of}
\keyword{overwrite}
\keyword{pdf}
\keyword{pdf("CCxxxTest.pdf")}
\keyword{setup}
\keyword{studio}
\keyword{that}
\keyword{the}
\keyword{this}
\keyword{to}
\keyword{use}
\keyword{using}
\keyword{way}
\keyword{write}

