% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ssLC}
\alias{ssLC}
\title{Return subString for each item in list}
\usage{
ssLC(I1, I2, line, sep)
}
\arguments{
\item{I1}{integer}

\item{I2}{integer}

\item{line}{string}

\item{sep}{string}
}
\description{
Return subString for each item in list
}

