% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoFRA.R
\name{summary.CompleteEnrichment}
\alias{summary.CompleteEnrichment}
\title{This function summarize and print the results from complete functional enrichment analysis}
\usage{
\method{summary}{CompleteEnrichment}(object, ...)
}
\arguments{
\item{object}{object from complete functional enrichment analysis}

\item{...}{list of additional arguments}
}
\description{
This function summarize and print the results from complete functional enrichment analysis
}
\keyword{"MCT","iNT","sNT","AllT,MCCTT,MTT,MCT,iNT,sNT"))}
\keyword{"MTT","MT","MT","MT","sN","sN","sN","sNT","sNT","sNT","iN","iN","iN","iNT","iNT","iNT"))}
\keyword{CC1=CoFRA::completeFunctionalRegulationAnalysis(iBAQ,Func[100:200],Fac,dfComp,NbackGround=142140)}
\keyword{Fac=factor(c("MCCTT","MCCTT","MCCTT","MCCT","MCCT","MCCT","MC","MC","MC","MCT","MCT","MCT","MTT","MTT",}
\keyword{Func=CoFRA::getFunctionalCategories("CC")}
\keyword{data(iBAQ)}
\keyword{dfComp=data.frame(Con=c("MCCT","MT","MC","iN","sN","AllC,MCCT,MT,MC,iN,sN"),Tre=c("MCCTT","MTT",}
\keyword{head(str(Func))}
\keyword{library(CoFRA)}
\keyword{summary}
\keyword{summary.CompleteEnrichment(CC1)}

