% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoFRA.R
\name{getFunctionalCategories}
\alias{getFunctionalCategories}
\title{get data frame with functional categories}
\usage{
getFunctionalCategories(func = "CC")
}
\arguments{
\item{func}{string ("BP","CC","MF")}
}
\description{
get data frame with functional categories
}
\examples{
library(CoFRA)
Acc=getFunctionalCategories(func="CC")
}
\keyword{Gene}
\keyword{ontology}

